# Backend/openapi.json
{"openapi":"3.1.0","info":{"title":"JumBah Backend API","description":"Backend API for JumBah travel application","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health":{"get":{"summary":"Health Check","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/locations/search":{"get":{"tags":["locations"],"summary":"Search Locations","description":"Search for locations using OpenStreetMap Nominatim API\n\n- **query**: The location name or address to search for\n- **limit**: Maximum number of results to return (1-50)","operationId":"search_locations_api_locations_search_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","description":"Location search query","title":"Query"},"description":"Location search query"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Maximum number of results","default":10,"title":"Limit"},"description":"Maximum number of results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/locations/geocode":{"get":{"tags":["locations"],"summary":"Geocode Location","description":"Geocode an address to get coordinates\n\n- **address**: The address to geocode\n- **limit**: Maximum number of results to return (1-10)","operationId":"geocode_location_api_locations_geocode_get","parameters":[{"name":"address","in":"query","required":true,"schema":{"type":"string","description":"Address to geocode","title":"Address"},"description":"Address to geocode"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":10,"minimum":1,"description":"Maximum number of results","default":1,"title":"Limit"},"description":"Maximum number of results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/locations/reverse-geocode":{"get":{"tags":["locations"],"summary":"Reverse Geocode","description":"Reverse geocode coordinates to get address information\n\n- **lat**: Latitude coordinate\n- **lon**: Longitude coordinate","operationId":"reverse_geocode_api_locations_reverse_geocode_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","description":"Latitude","title":"Lat"},"description":"Latitude"},{"name":"lon","in":"query","required":true,"schema":{"type":"number","description":"Longitude","title":"Lon"},"description":"Longitude"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/locations/nearby-pois":{"get":{"tags":["locations"],"summary":"Get Nearby Pois","description":"Get nearby points of interest around a location\n\n- **lat**: Latitude coordinate\n- **lon**: Longitude coordinate\n- **radius**: Search radius in meters (100-5000)\n- **poi_types**: Comma-separated list of POI types to search for","operationId":"get_nearby_pois_api_locations_nearby_pois_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","description":"Latitude","title":"Lat"},"description":"Latitude"},{"name":"lon","in":"query","required":true,"schema":{"type":"number","description":"Longitude","title":"Lon"},"description":"Longitude"},{"name":"radius","in":"query","required":false,"schema":{"type":"integer","maximum":5000,"minimum":100,"description":"Search radius in meters","default":1000,"title":"Radius"},"description":"Search radius in meters"},{"name":"poi_types","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Comma-separated POI types (restaurant,hotel,tourist,shop,amenity)","title":"Poi Types"},"description":"Comma-separated POI types (restaurant,hotel,tourist,shop,amenity)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/locations/route":{"get":{"tags":["locations"],"summary":"Get Route","description":"Get route between two points\n\n- **start_lat**: Starting point latitude\n- **start_lon**: Starting point longitude\n- **end_lat**: Destination latitude\n- **end_lon**: Destination longitude\n- **profile**: Transportation mode (driving, walking, cycling)","operationId":"get_route_api_locations_route_get","parameters":[{"name":"start_lat","in":"query","required":true,"schema":{"type":"number","description":"Start latitude","title":"Start Lat"},"description":"Start latitude"},{"name":"start_lon","in":"query","required":true,"schema":{"type":"number","description":"Start longitude","title":"Start Lon"},"description":"Start longitude"},{"name":"end_lat","in":"query","required":true,"schema":{"type":"number","description":"End latitude","title":"End Lat"},"description":"End latitude"},{"name":"end_lon","in":"query","required":true,"schema":{"type":"number","description":"End longitude","title":"End Lon"},"description":"End longitude"},{"name":"profile","in":"query","required":false,"schema":{"type":"string","description":"Routing profile (driving, walking, cycling)","default":"driving","title":"Profile"},"description":"Routing profile (driving, walking, cycling)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ai/generate-plan":{"post":{"tags":["ai-planner"],"summary":"Generate Travel Plan","description":"Generate a comprehensive travel plan using AI\n\n- **destination**: Travel destination\n- **duration**: Trip duration (e.g., \"3 days\", \"1 week\")\n- **budget**: Budget range (optional)\n- **preferences**: List of preferences (optional)\n- **user_id**: User identifier for chat history (optional)","operationId":"generate_travel_plan_api_ai_generate_plan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TravelPlanRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ai/chat":{"post":{"tags":["ai-planner"],"summary":"Chat With Ai","description":"Chat with AI for travel-related questions\n\n- **message**: User message/question\n- **user_id**: User identifier for chat history (optional)\n- **context**: Additional context for the conversation (optional)","operationId":"chat_with_ai_api_ai_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ai/history/{user_id}":{"get":{"tags":["ai-planner"],"summary":"Get Chat History","description":"Get chat history for a specific user\n\n- **user_id**: User identifier\n- **limit**: Maximum number of records to return (1-100)","operationId":"get_chat_history_api_ai_history__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Maximum number of records","default":50,"title":"Limit"},"description":"Maximum number of records"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ai/health":{"get":{"tags":["ai-planner"],"summary":"Ai Health Check","description":"Check if AI service is available","operationId":"ai_health_check_api_ai_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/weather/test":{"get":{"tags":["weather"],"summary":"Test Weather","description":"Test endpoint to verify weather router is working","operationId":"test_weather_api_weather_test_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/weather":{"get":{"tags":["weather"],"summary":"Get Weather","description":"Get current weather data for given coordinates","operationId":"get_weather_api_weather_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","title":"Lat"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","title":"Lon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/weather/location":{"get":{"tags":["weather"],"summary":"Get Weather By Location","description":"Get current weather data by location name (city, region, etc.)","operationId":"get_weather_by_location_api_weather_location_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id","default":"anonymous"},"context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context"}},"type":"object","required":["message"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TravelPlanRequest":{"properties":{"destination":{"type":"string","title":"Destination"},"duration":{"type":"string","title":"Duration"},"budget":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Budget"},"preferences":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferences"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id","default":"anonymous"}},"type":"object","required":["destination","duration"],"title":"TravelPlanRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}